416366: [CDO] Support implicit root class in server-side CDO queries https://bugs.eclipse.org/bugs/show_bug.cgi?id= 416366
Add a 'cdoImplicitRootClass' query option to the OCLQueryHandler,
similar to the existing 'cdoLazyExtents'.  The value is an EClass,
default is null.  Because the value is an EClass, it is not supported as
a variable that is passed through to the OCL expression parsing and
evaluation environments, so it is filtered out of the OCL environment.
The lazy-extent parameter also doesn't make sense in the OCL expression.
Change-Id: Ic4c40965fcce2dad092943a526b75a816421ea31