[418454] [Admin] Client API and UI for managing repositories in a server https://bugs.eclipse.org/bugs/show_bug.cgi?id=418454
Prompt the user to confirm deletion of a repository if any users are currently connected to it.  This uses a new reusable confirmation signal API, including confirmation providers, after the fashion of the authentication signal and credentials providers.

Also added the "Manage Security" action to repositories in the CDO Administration view.  This uses an alternative handler targeting an admin-repository, which opens a new session if an open session doesn't already exist that is logged in as the administrator.

Change-Id: Ib8d7e7d0237f9426158f5a9af57782b8515531cc
Also-by: Eike Stepper <stepper@esc-net.de>