[447593] Already in cache when calling twice "CDOTransaction.setBranch();"
Update DBUtil.close() to throw a DBException instead of returning a
Exception which will be ignored, when the
"-Dorg.eclipse.net4j.db.close.noisy=true" is specified.
Reapply partial fix of
https://bugs.eclipse.org/bugs/show_bug.cgi?id=370105 , only the
"readLists" part and not the "readValuesFromStatement" which is the
original cause of bug
https://bugs.eclipse.org/bugs/show_bug.cgi?id=406766 .
In addition update AllTests and GerritTests launch configuration to have
the noisy close enabled during tests execution.

Change-Id: I6ed6f2d6d0ead05873c37c43fa6f43b1b5095d2f
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=447593
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>