[580981] [DB] Add catalog parameter for DB meta data queries
For MariaDB, the 'catalog' is used as the top level namespace separator i.e. the database name.

This should be used from the connection (normally specified in the URL) as otherwise the queries return the metadata for all the tables in the database (and visible to the user) instead of only from the currently selected database

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580981
Change-Id: I0207f1c124d434e250fa6f626c09a29a4ba5fb3b
Signed-off-by: Eike Stepper <stepper@esc-net.de>